Many changes to compile plugins on windows with Visual Studio 2008 C++ Still need to finish GB_AmiMessage features
git-svn-id: file:///home/karl/SVN2/amilab.svn/code/trunk@1412 52b0d089-85ac-4fb5-acf5-4e3172ea7aaf

+++ src/Algorithms/Vaisseaux/IsoLigne2.cpp
@@ -69,9 +69,9 @@ int IsoLigne2 :: RechercheLignes( int num_face, num_valeurs val,
     LigneFace* lignes_zero;
     int     i,j;
     unsigned char    sommet_dans_liste[4];
-    double* valeurs=NULL;
-    char        message[100];
-    float        a,b,c,d,S,f;
+    double*          valeurs=NULL;
+    char             message[100];
+    double           a,b,c,d,S,f;
 
   Pour(i,0,3)
     sommet_dans_liste[i] = false;
