+++ src/Algorithms/myVTK4/PrincipalCurvatures.cxx
@@ -103,7 +103,7 @@ int CurvaturasPrincipales(float H[3][3],
   /* NORMALIZO EL VECTOR GRADIENTE Y LO GUARDO EN p1[3]*/
   pnorm=sqrt((double) pnorm);
   for(i=0;i<3;i++)
-    p1[i]=p[i]/pnorm;
+    p1[i]= (float) p[i]/pnorm;
 
 
  /* CALCULO CALCULO A=(Id-pxp/p^2)H(Id-pxp/p^2) (VERSION OPTIMIZADA)*/
@@ -169,7 +169,7 @@ int CurvaturasPrincipales(float H[3][3],
   d=sqrt((c-a)*(c-a)+4*b*b);
   e=c+a;
   if(e>0){
-    *lmax=0.5*(e+d);
+    *lmax= (float) (0.5*(e+d));
     if( autovector0((a-*lmax),b,(c-*lmax),paso)==-1){  
       paso[0]=0;
       paso[1]=1;
@@ -180,7 +180,7 @@ int CurvaturasPrincipales(float H[3][3],
     for(i=0;i<3;i++)
       for(j=0;j<3;j++)
         vmax[i]+=P[i][j]*paso[j];
-    *lmin=0.5*(e-d);
+    *lmin= (float) (0.5*(e-d));
     if( autovector0((a-*lmin),b,(c-*lmin),paso)==-1){  
       paso[0]=0;
       paso[1]=0;
@@ -193,7 +193,7 @@ int CurvaturasPrincipales(float H[3][3],
         vmin[i]+=P[i][j]*paso[j];
   }
   else{
-    *lmin=0.5*(e+d);
+    *lmin= (float) (0.5*(e+d));
     if( autovector0((a-*lmin),b,(c-*lmin),paso)==-1){  
       paso[0]=0;
       paso[1]=0;
@@ -218,8 +218,8 @@ int CurvaturasPrincipales(float H[3][3],
   }
 
   /* NORMALIZO LOS AUTOVALORES PARA OBTENER LAS CURVATURAS PRINCIPALES */
-  *lmax/=-pnorm;
-  *lmin/=-pnorm;
+  *lmax/=(float)-pnorm;
+  *lmin/=(float)-pnorm;
 
   //  printf("vmin %f %f %f \n", vmin[0], vmin[1], vmin[2]);
  
@@ -257,7 +257,7 @@ int CurvaturasPrincipales(float H[3][3],
   /* NORMALIZO EL VECTOR GRADIENTE Y LO GUARDO EN p1[3]*/
   pnorm=sqrt((double) pnorm);
   for(i=0;i<3;i++)
-    p1[i]=p[i]/pnorm;
+    p1[i]= (float) (p[i]/pnorm);
 
 
  /* CALCULO CALCULO A=(Id-pxp/p^2)H(Id-pxp/p^2) (VERSION OPTIMIZADA)*/
@@ -271,7 +271,7 @@ int CurvaturasPrincipales(float H[3][3],
 
   for(i=0;i<3;i++)
     for(j=i;j<3;j++)
-      A[i][j]=H[i][j]-paso[i]*p1[j]+(a*p1[j]-paso[j])*p1[i]; 
+      A[i][j]=(float) (H[i][j]-paso[i]*p1[j]+(a*p1[j]-paso[j])*p1[i]); 
 
   for(i=0;i<3;i++)
     for(j=0;j<i;j++)
@@ -290,8 +290,8 @@ int CurvaturasPrincipales(float H[3][3],
   p2[k2]=0;
   p2norm=p2[k]*p2[k]+p2[k1]*p2[k1];
   p2norm=sqrt(p2norm);
-  p2[k]/=p2norm; 
-  p2[k1]/=p2norm;  
+  p2[k]/=(float)p2norm; 
+  p2[k1]/=(float)p2norm;  
 
   p3[k]=-p1[k2]*p2[k1];
   p3[k1]=p1[k2]*p2[k];
@@ -323,17 +323,17 @@ int CurvaturasPrincipales(float H[3][3],
   d=sqrt((c-a)*(c-a)+4*b*b);
   e=c+a;
   if(e>0){
-    *lmax=0.5*(e+d);
-    *lmin=0.5*(e-d);
+    *lmax=(float)(0.5*(e+d));
+    *lmin=(float)(0.5*(e-d));
   }
   else{
-    *lmin=0.5*(e+d);
-    *lmax=0.5*(e-d);
+    *lmin=(float)(0.5*(e+d));
+    *lmax=(float)(0.5*(e-d));
   }
 
   /* NORMALIZO LOS AUTOVALORES PARA OBTENER LAS CURVATURAS PRINCIPALES */
-  *lmax/=-pnorm;
-  *lmin/=-pnorm;
+  *lmax/=(float)-pnorm;
+  *lmin/=(float)-pnorm;
 
   return(1);
